refactor: use default context to call agent if not provided
