+++ examples/workflow-group-chat/index.ts
@@ -19,9 +19,13 @@ import { z } from "zod";
 const { OPENAI_API_KEY } = process.env;
 assert(OPENAI_API_KEY, "Please set the OPENAI_API_KEY environment variable");
 
-const gpt = new OpenAIChatModel({
+const model = new OpenAIChatModel({
   apiKey: OPENAI_API_KEY,
   model: "gpt-4o",
+});
+
+const engine = new ExecutionEngine({
+  model,
 });
 
 const DEFAULT_TOPIC = "DEFAULT_TOPIC";
@@ -83,6 +87,7 @@ Make sure the images have consistent characters and style.`,
 let isFirstQuestion = true;
 
 const user = UserAgent.from({
+  context: engine.newContext(),
   name: "user",
   description: "User for providing final approval",
   publishTopic: DEFAULT_TOPIC,
@@ -136,10 +141,7 @@ const manager = AIAgent.from({
   }),
 });
 
-const engine = new ExecutionEngine({
-  model: gpt,
-  agents: [user, manager, writer, editor, illustrator],
-});
+engine.addAgent(user, writer, editor, illustrator, manager);
 
 engine.subscribe(DEFAULT_TOPIC, (message) => {
   console.log(
